Simplify the ContentIterator by making it synchronous
