+++ Sources/Shared/Toolkit/Extensions/Optional.swift
@@ -28,5 +28,11 @@ extension Optional {
         }
         return value
     }
-    
+
+    /// Returns the wrapped value and modify the variable to be nil.
+    mutating func pop() -> Wrapped? {
+        let res = self
+        self = nil
+        return res
+    }
 }
